Test skipping code owner config validation when rebasing via API
If a patch set is created via the Rebase REST
endpoint the code owner config file validation validates any OWNERS file
that is touched in the commit. This validation fails if such an OWNERS
file contains conflicts (conflicts are possible if
rebaseInput.allowConflicts = true was set), since the conflict
markers in the file will be detected as invalid content.

Change If025ac253 in Gerrit core allows to set validation options in the
rebase input that are forwarded to the CommitValidationListeners. This
makes it possible to skip the code owner config validation when creating
a patch set via the Rebase REST endpoint, by specifying
code-owners~skip-validation=true as validation option.

This change adds a test to verify that skipping the code owner config
validation by setting code-owners~skip-validation=true in the rebase
input works.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I852391f4d3ba823056a603052dcb259bad5b97bc