Validate OWNERS files on branch creation
With this change it's possible to enable validation of code owner config
files on branch creation (off by default). Enabling this validation is
rather expensive and makes branch creations slower (in average a latency
increase of 10s to 20s is expected).

If the validation is enabled and a new branch is created, all code owner
config files that are contained in the initial commit are newly
validated, even if the branch is created for a commit that already
exists in the repository. This makes sense since the validation
configuration of the new branch may differ from the branch that already
contains the commit. Also, if code owner config files are imported from
other projects without specifying a branch, these imports may be broken
in the context of the new branch (because the import assumes the same
branch as the importing code owner config and that branch may not exist
in the other project) and hence should be flagged on branch creation.

Skipping the validation is possible by specifying the
code-owners~skip-validation=true validation option.

Change-Id: I57db5b4735eff591c0de3e8c1cc22b7dc27226e3
Signed-off-by: Edwin Kempin <ekempin@google.com>