Validate OWNERS files on branch creation
When a new branch is created, all code owner config files that are
contained in the initial commit are newly validated now, even if the
branch is created for a commit that already exists in the repository.
This is necessary since the validation configuration of the new branch
may differ from the branch that already contains the commit. Also, if
code owner config files are imported from other projects without
specifying a branch, these imports may be broken in the context of the
new branch (because the import assumes the same branch as the importing
code owner config and that branch may not exist in the other project)
and hence should be flagged on branch creation.

Skipping the validation is possible by specifying the
code-owners~skip-validation=true validation option.

In addition it's possible to disable the validation via the plugin
configuration.

Change-Id: I57db5b4735eff591c0de3e8c1cc22b7dc27226e3
Signed-off-by: Edwin Kempin <ekempin@google.com>