README: Include links to properly rendered plugin documentation
Explain that the plugin documentation only renders correctly when the
plugin is installed in Gerrit and the documentation is accessed via
https://<gerrit-host>/plugins/code-owners/Documentation/index.html.

This is because Gerrit generates html files out of the Markdown files
and those are referenced as target for links. If the documentation is
viewed in Gitiles the html files are not present and the links cannot be
resolved. In addition the plugin documentation contains relative links
to the Gerrit core documentation that can also not be resolved in
Gitiles. Finally Gerrit also takes care to replace macros, which also
doesn't happen when the documentation is viewed in Gitiles.

To allow users read the documentation without installing the plugin, we
include links to the code-owners plugin documentation that is hosted on
the android host, where it is properly rendered.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I7a51ea20a171e43fff5c140c5933aea64a7ba4cd