Validate code owner config files in parallel
If a commit adds/modifies many OWNERS files at once, it's better to
validate them in parallel rather than sequentially.

To do the parallel execution we use the existing fan out executor.

Doing the validations of OWNERS files in parallel gets more important if
we start to validate OWNERS files on branch creation, as in this case we
need to validate all existing OWNERS files (that are contained in the
commit on which the branch is created) and not only newly added and
modified OWNERS files. Validating OWNERS files on branch creation will
be added in a follow-up change.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I443062ca57573b41b050c04106461765912f27cc