Warn if code-owners plugin configuration is added in project.config
The configuration of the code-owners plugin is read from
* gerrit.config (section 'plugin.code-owners')
* code-owners.config in refs/meta/config (mostly section 'codeOwners')

Sometimes it happens that users wrongly add code-owners plugin
configuration in project.config and then don't understand why the
configuration doesn't work.

Print out a warning if code-owners plugin configuration is added in
project.config so that users get to know that this is the wrong place to
configure the code-owners plugin.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ice385139c41cae421297f12895cdd1a6efff8b93