CodeOwnerApprovalCheck: Factor out computation of input data
The CodeOwnerApprovalCheck class is quite large which makes the class
difficult to read. To make it smaller we factor out the computation of
all input data that this required for checking the code owner approvals
of a change. The input data is now represented by the new
CodeOwnerApprovalCheckInput class which is an immutable AutoValue class.
The loading of the input is done by the
CodeOwnerApprovalCheckInput.Loader class (e.g. this class computes the
relevant reviewers and approvers, loads the global code owners and
fallback code owners from the config etc.). The Loader class needs to
get some classes injected which is why we have a Factory to create it.

Having a class that represents all inputs, make it easier for
CodeOwnerApprovalCheck to pass down the input through the methods that
do the computation (just pass 1 parameter instead of n parameters).

In addition to checking code owner approvals on a change
CodeOwnerApprovalCheck also computes owned paths for a set of users.
This is achieved by using the same algorithm, but with specially
constructed inputs. By having the CodeOwnerApprovalCheckInput class we
can make the creation of the input for this purpose more explicit.

In future we need additional inputs for the code owner approval check
(e.g. approvals on previous patch sets for making code owner approvals
sticky on file level). Adding these additional inputs is easier now that
we have CodeOwnerApprovalCheckInput, otherwise we would need to add even
more logic to the already large CodeOwnerApprovalCheck class.

This change is a pure refactoring. This means no behavior is changed.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I298f8e3b6acd7afdec900e1c1fde6d1f26ed1a06