Add unit tests for CodeOwnerApprovalCheckInput
Now that the logic for computing the input for CodeOwnerApprovalCheck is
in a separate class we can easily unit test it.

We already have integration tests that cover this logic, but the unit
tests are easier to debug if anything gets broken.

This change doesn't add tests for CodeOwnerApprovalCheckInputs that are
created by the createForComputingOwnedPaths method yet, we may add those
later.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I9e63ce035c0212454372435db9edf3c565069c8b