Show files as unknown when their code-owner status is not present
The absence of files in code-owners needn't be an error as originally
coded. It could be due to the fact that the files are not part of the
change list (if a file is being commented on that's not part of the
change). Or it could be files from a merge commit.

Instead show their status as not needing code-owners approval.

Screenshot: (Before) https://imgur.com/a/qf47aAj, (After) https://imgur.com/a/J3PuGGU
Google-Bug-Id: b/228404313, b/253451848
Change-Id: I73d1c2b7c06714d1ac438a66bbf6fc9389fd2238