Support showing owner avatars.
When the current user does not own a file, we show a stack of at most 3
avatars of reviewers that are owners for this user. If the current user
does own the file, the shield icon is shown instead.

To fetch this information, `owned_paths` is passed a new parameter
`check_reviewers` which was introduced in Change 348257.

Google-Bug-Id: b/200008775
Screenshot: https://imgur.com/a/uLzzmpP
Change-Id: Ifb841a7316ada294c26e9c76221b4581ccb20d94