Compute owners for file-deletions correctly.
Gerrit FE returns 'path' not 'oldPath' for deleted files. However,
code-owners plugin computes deleted files as 'oldPath' not 'path'.  To
get around this difference in API, we check if oldPath is not set. In
that case, we also check if path exists in the oldPaths returned by the
code-owners plugin

Google-Bug-Id: b/200008775
Change-Id: I268397ccf1ed5d89cfe03b64b42b917c67c80233