Improve error message when trying to specify code owners by username
Code owners in OWNERS files need to be specified by email.

It's a common mistake by our users to specify usernames instead (e.g.
'foo' instead of 'foo@google.com').

So far code owners that were specifed by username were rejected with a
general "invalid line: ..." error and users had trouble spotting what
was wrong.

With this change we return a better error message now: "code owner ...
in ... cannot be resolved (must be an email)"

The generic error was thrown from the parser that parses OWNERS files.
We relaxed the matching for code owners there, so that code owners that
are specified by username do not cause a parsing error. These code
owners are not resolvable and hence parsing them is OK. In
CodeOwnerConfigValidator we add a special check for code owners that
are usernames and reject them with the better error message.

Since "INVALID" as a line in an OWNERS file is no longer causing a
parsing error we have to adapt the tests that used this string to
trigger a parsing error. Now we use "@INVALID" instead.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I04efc5028a629e017614cc8c1963f586ec4b5d5b