Remove the bug-report link from code-owners
The bug-report link had been added to ease the feedback process during
the code owners rollout, but since code owners is pretty mature by now
it's no longer needed.

Change I1027d4aa7 already removed the link in the master branch, but the
code in the stable-3.3 looks slightly different, so we can't cherry-pick
that change.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia226ec473e02c3fd513f37374ef30f7afc0ceded