RenameEmail: Preserve files which are not updated
The RenameEmail REST endpoint allows to update emails in code owner
config files. The commit that was created when this REST endpoint was
invoked only contained the modified code owner config files and all
other files, non code owner config files and code owner config files
that didn't contain the email, were deleted. This was because we created
a new tree that we populated with the modifications instead of using the
existing tree as the base.

This being broken means that this REST endpoint hasn't been used so far.

Bug: Issue 16781
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Id396ff9c7dacc210d761baee855d3b2c51f64bac
Signed-off-by: Edwin Kempin <ekempin@google.com>