Cache missing projects of unresolved imports for the time of the request
ProjectCache does not cache non-existing projects, but each time a
non-existing project is requested it tries to load it from the storage.
Within Google this load can take ~30ms which is not much, but it adds up
if a non-existing project is requested multiple times. For the
code-owners plugin this can be a problem if OWNERS files that are
relevant for many files in a change (e.g. the root OWNERS file) contain
unreolvable imports that point to non-existing projects. For each file
for which an OWNERS files is relevant it is evaluated once in
PathCodeOwners. Evaluating the OWNERS file once per file is needed
because depending on the file different parts of the OWNERS file apply.
Those imports that apply to the file are resolved during the evaluation.
Note: Each OWNERS file is loaded only once, but it is
evaluated once for each file to which it applies. This means in the
worst case a non-existing project is looked up by PathCodeOwners from
the project cache once per file in the change. In addiition
UnresolvedImportFormatter needs to load the code owners project config
in order to format unresolved imports (it needs to know the default file
name for OWNERS files which depends on the used code owners backend).
Also for this project is retrieved once per file from the cache, which
means in total an unresolvable import in the root OWNERS file that
points to a non-existing project can cause the non-existing project to
be retrieved 2 * number of files in the change (e.g. when checking if
the change is submittable).

To avoid repeated lookups of a non-existing project that each take 30ms,
we now cache the non-existing projects for the time of the request. This
means each non-existing project is looked up from ProjectCache only
once per request now. This is implemented by a per thread project cache
that caches missing projects.

Bug: Google b/284533026
Change-Id: Ia776041622eac9d2ad1f2bc0f96692a5d567651a
Signed-off-by: Edwin Kempin <ekempin@google.com>