Don't show config validation hints when code-owners is disabled
These hints don't provide a lot of value and are especially unwanted
when a server has lot of projects and code-owners is enabled only on
a handful of them.

Log a message at fine level so that it's visible from traces that the
code owners validation is skipped because the code owners functionality
is disabled.

Change-Id: Ieb0033351554ed339cba484daf7b7ca6b0614d93