Explain on the about page when the code-owners plugin should/shouldn't be used
This is already explained in the how to use guide, but users often miss
this information there. The about page is the entry page of the
code-owners documentation (it is integrated as the first section into
the index page) and what users that are interested in using the
code-owners plugin read first, hence users are unlikely to miss this
information here.

Change-Id: Ie3800fd96eadcbcaa97f610398f911b4a1db25b3
Signed-off-by: Edwin Kempin <ekempin@google.com>