Show that CodeOwnersSubmitRule is run twice when a Code-Owners SR is configured
Add a test to show that the CodeOwnersSubmitRule is executed twice when
a Code-Owners submit requirement is configured, once because the
CodeOwnerSubmitRule implements the SubmitRule extension point and once
because there is submit requirement configured that uses the
CodeOwnerApprovalPredicate.

CodeOwnerApprovalPredicate was added by change I391a0d7fd as part of
replacing submit rules with submit requirements, but then submit rules
were not migrated to submit requirements.

Bug: Google b/325070428
Change-Id: I07f9ed1f0dff448b5b51d9dd232184648e88652a
Signed-off-by: Edwin Kempin <ekempin@google.com>
Reviewed-on: https://gerrit-review.googlesource.com/c/plugins/code-owners/+/412200
Tested-by: Zuul <zuul-63@gerritcodereview-ci.iam.gserviceaccount.com>
Reviewed-by: Patrick Hiesel <hiesel@google.com>