Clarify that being able to approve changes doesn't make a user a code owner
Code owner approvals are applied by voting on a label. Which label
counts as a code owner approval is configurable. Sometimes users think
that being able to vote on that label makes them a code owner. Having
this permission is not sufficient, but to be considered as a code owner
users must be listed as a code owner in an OWNERS file. Add a note about
this to the documentation.

Release-Notes: skip
Change-Id: I39df35bb14ae81b6f6f2af69ed146a943924041d
Signed-off-by: Edwin Kempin <ekempin@google.com>